unplayed/requirements.txt
ag-ui-langgraph>=0.0.30
copilotkit>=0.1.43
langchain-ollama>=0.3.0
langchain-openai>=0.3.0
langchain-core>=1.2.10,<2.0.0
langchain>=1.2.10,<2.0.0
langgraph>=1.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
python-dotenv>=1.0.0
